register_Login

+++ BankSystemProject/Models/LoanApplication.cs
@@ -5,7 +5,7 @@
         public int LoanApplicationId { get; set; }
         public int CustomerAccountId { get; set; }
         public int LoanTypeId { get; set; }
-        public decimal AmountRequested { get; set; }
+        public double AmountRequested { get; set; }
         public DateTime ApplicationDate { get; set; }
         public string Status { get; set; }
         public DateTime? ApprovalDate { get; set; }
